fix(ts): clean up adapter interfaces
