+++ types/adapters.d.ts
@@ -1,245 +1,121 @@
 import { AppOptions } from "./internals"
-import { ConnectionOptions, EntitySchema } from "typeorm"
-import { User } from "."
-import { AppProvider } from "./providers"
+import { User, Profile, Session } from "."
+import { EmailConfig, SendVerificationRequest } from "./providers"
 
-export interface Profile {
-  id: string
-  name: string
-  email: string | null
-  image?: string | null
-}
-
-export interface Session {
-  userId: string | number | object
-  expires: Date
-  sessionToken: string
-  accessToken: string
-}
+/** Legacy */
+export * from "internals/legacy-adapters"
 
 export interface VerificationRequest {
+  id: string
   identifier: string
   token: string
   expires: Date
-}
-
-export interface SendVerificationRequestParams {
-  identifier: string
-  url: string
-  token: string
-  baseUrl: string
-  provider: AppProvider
-}
-
-export type EmailAppProvider = AppProvider & {
-  sendVerificationRequest: (
-    params: SendVerificationRequestParams
-  ) => Promise<void>
-  maxAge: number | undefined
-}
-
-export interface AdapterInstance<
-  TUser,
-  TProfile,
-  TSession,
-  TVerificationRequest
-> {
-  createUser: (profile: TProfile) => Promise<TUser>
-  getUser: (id: string) => Promise<TUser | null>
-  getUserByEmail: (email: string) => Promise<TUser | null>
-  getUserByProviderAccountId: (
-    providerId: string,
-    providerAccountId: string
-  ) => Promise<TUser | null>
-  updateUser: (user: TUser) => Promise<TUser>
-  linkAccount: (
-    userId: string,
-    providerId: string,
-    providerType: string,
-    providerAccountId: string,
-    refreshToken: string,
-    accessToken: string,
-    accessTokenExpires: number
-  ) => Promise<void>
-  createSession: (user: TUser) => Promise<TSession>
-  getSession: (sessionToken: string) => Promise<TSession | null>
-  updateSession: (session: TSession, force?: boolean) => Promise<TSession>
-  deleteSession: (sessionToken: string) => Promise<void>
-  createVerificationRequest?: (
-    email: string,
-    url: string,
-    token: string,
-    secret: string,
-    provider: EmailAppProvider,
-    options: AppOptions
-  ) => Promise<TVerificationRequest>
-  getVerificationRequest?: (
-    email: string,
-    verificationToken: string,
-    secret: string,
-    provider: AppProvider
-  ) => Promise<TVerificationRequest | null>
-  deleteVerificationRequest?: (
-    email: string,
-    verificationToken: string,
-    secret: string,
-    provider: AppProvider
-  ) => Promise<void>
-}
-
-interface Adapter<
-  TUser extends User = any,
-  TProfile extends Profile = any,
-  TSession extends Session = any,
-  TVerificationRequest extends VerificationRequest = any
-> {
-  getAdapter: (
-    appOptions: AppOptions
-  ) => Promise<AdapterInstance<TUser, TProfile, TSession, TVerificationRequest>>
-}
-
-type Schema<T = any> = EntitySchema<T>["options"]
-
-interface BuiltInAdapters {
-  Default: TypeORMAdapter["Adapter"]
-  TypeORM: TypeORMAdapter
-  Prisma: PrismaAdapter
 }
 
 /**
- * TODO: fix auto-type schema
+ * Using a custom adapter you can connect to any database backend or even several different databases.
+ * Custom adapters created and maintained by our community can be found in the adapters repository.
+ * Feel free to add a custom adapter from your project to the repository,
+ * or even become a maintainer of a certain adapter.
+ * Custom adapters can still be created and used in a project without being added to the repository.
+ *
+ * [Community adapters](https://github.com/nextauthjs/adapters) |
+ * [Create a custom adapter](https://next-auth.js.org/tutorials/creating-a-database-adapter)
  */
-
-interface TypeORMAdapter<
-  A extends TypeORMAccountModel = any,
-  U extends TypeORMUserModel = any,
-  S extends TypeORMSessionModel = any,
-  VR extends TypeORMVerificationRequestModel = any
-> {
-  Adapter: (
-    typeOrmConfig: ConnectionOptions,
-    options?: {
-      models?: {
-        Account?: {
-          model: A
-          schema: Schema<A>
-        }
-        User?: {
-          model: U
-          schema: Schema<U>
-        }
-        Session?: {
-          model: S
-          schema: Schema<S>
-        }
-        VerificationRequest?: {
-          model: VR
-          schema: Schema<VR>
-        }
-      }
-    }
-  ) => Adapter<U, Profile, S, VR>
-  Models: {
-    Account: {
-      model: TypeORMAccountModel
-      schema: Schema<TypeORMAccountModel>
-    }
-    User: {
-      model: TypeORMUserModel
-      schema: Schema<TypeORMUserModel>
-    }
-    Session: {
-      model: TypeORMSessionModel
-      schema: Schema<TypeORMSessionModel>
-    }
-    VerificationRequest: {
-      model: TypeORMVerificationRequestModel
-      schema: Schema<TypeORMVerificationRequestModel>
-    }
-  }
-}
-
-interface PrismaAdapter {
-  Adapter: (config: {
-    prisma: any
-    modelMapping?: {
-      User: string
-      Account: string
-      Session: string
-      VerificationRequest: string
-    }
-  }) => Adapter
-}
-
-declare class TypeORMAccountModel {
-  compoundId: string
-  userId: number
-  providerType: string
-  providerId: string
-  providerAccountId: string
-  refreshToken?: string
-  accessToken?: string
-  accessTokenExpires?: Date
-
-  constructor(
-    userId: number,
+export interface AdapterInstance<U = User, P = Profile, S = Session> {
+  createUser(profile: P): Promise<U>
+  getUser(id: string): Promise<U | null>
+  getUserByEmail(email: string): Promise<U | null>
+  getUserByProviderAccountId(
+    providerId: string,
+    providerAccountId: string
+  ): Promise<U | null>
+  updateUser(user: U): Promise<U>
+  /** @todo Implement */
+  deleteUser?(userId: string): Promise<void>
+  linkAccount(
+    userId: string,
     providerId: string,
     providerType: string,
     providerAccountId: string,
     refreshToken?: string,
     accessToken?: string,
-    accessTokenExpires?: Date
-  )
+    accessTokenExpires?: null
+  ): Promise<void>
+  /** @todo Implement */
+  unlinkAccont?(
+    userId: string,
+    providerId: string,
+    providerAccountId: string
+  ): Promise<void>
+  createSession(user: U): Promise<S>
+  getSession(sessionToken: string): Promise<S | null>
+  updateSession(session: S, force?: boolean): Promise<S>
+  deleteSession(sessionToken: string): Promise<void>
+  createVerificationRequest?: SendVerificationRequest
+  getVerificationRequest?(
+    identifier: string,
+    verificationToken: string,
+    secret: string,
+    provider: EmailConfig
+  ): Promise<VerificationRequest | null>
+  deleteVerificationRequest?(
+    identifier: string,
+    verificationToken: string,
+    secret: string,
+    provider: EmailConfig
+  ): Promise<void>
 }
 
-declare class TypeORMUserModel implements User {
-  name?: string
-  email?: string
-  image?: string
-  emailVerified?: Date
-
-  constructor(
-    name?: string,
-    email?: string,
-    image?: string,
-    emailVerified?: Date
-  )
-  [x: string]: unknown
-}
-
-declare class TypeORMSessionModel implements Session {
-  userId: number
-  expires: Date
-  sessionToken: string
-  accessToken: string
-
-  constructor(
-    userId: number,
-    expires: Date,
-    sessionToken?: string,
-    accessToken?: string
-  )
-}
-
-declare class TypeORMVerificationRequestModel implements VerificationRequest {
-  identifier: string
-  token: string
-  expires: Date
-
-  constructor(identifier: string, token: string, expires: Date)
-}
-
-declare const Adapters: BuiltInAdapters
-
-export default Adapters
-
-export {
-  Adapter,
-  BuiltInAdapters as Adapters,
-  TypeORMAdapter,
-  TypeORMAccountModel,
-  TypeORMUserModel,
-  TypeORMSessionModel,
-  TypeORMVerificationRequestModel,
-  PrismaAdapter,
+/**
+ * From an implementation perspective, an adapter in NextAuth.js is a function
+ * which returns an async `getAdapter()` method, which in turn returns a list of functions
+ * used to handle operations such as creating user, linking a user
+ * and an OAuth account or handling reading and writing sessions.
+ *
+ * It uses this approach to allow database connection logic to live in the `getAdapter()` method.
+ * By calling the function just before an action needs to happen,
+ * it is possible to check database connection status and handle connecting / reconnecting
+ * to a database as required.
+ *
+ * **Required methods**
+ *
+ * _(These methods are required for all sign in flows)_
+ * - `createUser`
+ * - `getUser`
+ * - `getUserByEmail`
+ * - `getUserByProviderAccountId`
+ * - `linkAccount`
+ * - `createSession`
+ * - `getSession`
+ * - `updateSession`
+ * - `deleteSession`
+ * - `updateUser`
+ *
+ * _(Required to support email / passwordless sign in)_
+ *
+ * - `createVerificationRequest`
+ * - `getVerificationRequest`
+ * - `deleteVerificationRequest`
+ *
+ * **Unimplemented methods**
+ *
+ * _(These methods will be required in a future release, but are not yet invoked)_
+ * - `deleteUser`
+ * - `unlinkAccount`
+ *
+ * [Community adapters](https://github.com/nextauthjs/adapters) |
+ * [Create a custom adapter](https://next-auth.js.org/tutorials/creating-a-database-adapter)
+ */
+export type Adapter<
+  C = Record<string, unknown>,
+  O = Record<string, unknown>,
+  U = User,
+  P = Profile,
+  S = Session
+> = (
+  config: C,
+  options?: O
+) => {
+  getAdapter(appOptions: AppOptions): Promise<AdapterInstance<U, P, S>>
 }
